Added initial structure for a svglogger.
I often find myself wondering how certain Polygons, Polylines or
VariableWidthPaths (libArachne) are constructed. So I intend to build
a simple logger with multiple sinks where an arbitrary geometric type
can be passed in.

Each sink will be a SVG 2.0 file, which has support for multiple layers.
defined by the args... a Legend for easy reading and different line
types, colors and Area's with support of transparency.

It is intended to leave the logs scattered in strategic places in the
code base because they will be setup in such a way that they point to
void objects when compiled as Release.