Added a cmake Conan wrapper
The wrapper can be enabled by the user by parsing -DENABLE_CONAN=ON
at the cmake args. It will download the 0.15 wrapper from
https://github.com/conan-io/cmake-conan and place it in the
CMAKE_BINARY_DIR. It will then automatically add the bincrafters remote
which is remote repository with a lot of high quality Conan packages.

Cmake will create Find scripts for all dependencies and place these
scripts in the CMAKE_BINARY_DIR which is added to CMAKE_MODULE_PATH
This means that current way of working in the using the find_package
method still is valid. So using Conan is optional and it doesn't break
current behaviour. If you choose to use Conan the dependencies it
provides will take precedence.

This commit allows both stb and clipper to be handled by Conan. If these
were build already They will be reused, it will find the package in the
remote and either download the source and the binaries (if available) or
it will build the binaries from source.

My goal is to have more control over the way we handle internal and
external dependencies, speed up the build process and make it easier for
 developers to build from source.

Note:
- To actually us this Conan needs to be downloaded.
See https://conan.io/downloads.html
- at the current time the stb dependency handling differs from our old
 way. I still need to work on that but for now I added a preprocessor
 flag to deal with the difference in including the header.